refactor: extract handleNextAction from tools.ts to next-action.ts
