+++ plugins/exarchos/servers/exarchos-mcp/src/index.ts
@@ -11,12 +11,12 @@ import {
   handleSet,
   handleSummary,
   handleReconcile,
-  handleNextAction,
   handleTransitions,
   handleCancel,
   handleCheckpoint,
 } from './workflow/tools.js';
 import type { ToolResult } from './workflow/tools.js';
+import { handleNextAction } from './workflow/next-action.js';
 import {
   handleEventAppend,
   handleEventQuery,
